feat(cowfi): upgrade learn pages
