+++ apps/cow-fi/app/learn/articles/[[...page]]/layout.tsx
@@ -1,0 +1,9 @@
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'All articles',
+}
+
+export default function LayoutPage({ children }: { children: React.ReactNode }) {
+  return children
+}
