+++ apps/cow-fi/app/learn/topics/layout.tsx
@@ -1,0 +1,9 @@
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Topics',
+}
+
+export default function LayoutPage({ children }: { children: React.ReactNode }) {
+  return children
+}
