--- libs/core/src/cms/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useCmsSolversInfo'
-export * from './useSolversInfo'
