+++ libs/core/src/cms/index.ts
@@ -1,5 +1,3 @@
-export * from './hooks'
 export * from './types'
 export * from './state'
 export * from './utils'
-export * from './updaters'
