--- libs/core/src/cms/updaters/index.ts
@@ -1,1 +1,0 @@
-export * from './SolversInfoUpdater'
